Implement MVC & migrate from session to JWT
We implement MVC in case we need to make a bunch of route that maybe will be to much to store on the app.js file
We also consider to migrate from session to JWT because it is more convenient to store a user token on mobile local storage

Added some security consideration by adding salt & hashing functionality with bcrypt, because previosuly it is done automaticyly by passport.
Added some more error handling & validator in case the user authentication data is not valid.